First CLI flows working.

+++ salsa/gg/scripts/build_and_deploy.sh
@@ -1,4 +1,10 @@
-bazel build //salsa/gg:gg_cli --build_python_zip
+bzl test //salsa/gg:gg_test
+if [ $? -ne 0 ]
+then
+    exit 1
+fi
+
+bzl build //salsa/gg:gg_cli --build_python_zip
 rm -f ~/bin/gg
 
 echo '#!/usr/bin/env python' > ~/bin/gg
